refactor(apps/client): HASHI-169 예약 상세 타입 의존성 정리
예약 상세 view model이 UI 컴포넌트의 타입을 직접 참조하지 않도록 page-local 타입 경계를 분리한다.

컴포넌트와 util이 중립적인 화면 데이터 타입을 공유해 의존 방향을 단순화하고 문자열 표시 계약을 명확히 한다.
