+++ apps/client/src/pages/reservationDetail/ReservationDetailPage.spec.md
@@ -56,6 +56,7 @@
 - [ ] fixed Header와 fixed ActionBar는 z-index 토큰을 사용합니다.
 - [ ] 취소된 예약(`reservationStatus: CANCELED`)은 URL 직접 접근으로도 상세 화면을 표시하지 않고 `NotFoundPage`를 표시합니다.
 - [ ] 예약 취소 요청 중에는 취소 확인 버튼과 닫기 버튼을 비활성화해 중복 요청과 중간 닫기를 방지합니다.
+- [ ] `createReservationDetailViewModel`과 UI 컴포넌트가 공유하는 화면 데이터 타입은 page-local `types.ts`에서 관리하며, util은 `components/`를 import하지 않습니다.
 
 ## Data Dependencies
 
@@ -204,6 +205,10 @@ ReservationDetailPage
 - page-local util:
   - `createReservationDetailViewModel`
   - `reservationDetailPolicy`
+- page-local type:
+  - `ReservationProgressStatus`
+  - `ReservationProgressStep`
+  - `ReservationReceiptInfoItem`
 - page-local constants:
   - `reservationNotices`
 - icon:
