+++ apps/client/src/pages/reservationDetail/types.ts
@@ -1,0 +1,14 @@
+export type ReservationProgressStatus = 'completed' | 'current' | 'pending'
+
+export interface ReservationProgressStep {
+  id: string
+  title: string
+  description: string
+  requestedAt?: string
+  status: ReservationProgressStatus
+}
+
+export interface ReservationReceiptInfoItem {
+  label: string
+  value: string
+}
